project_1/requirements.txt
pytesseract
Pillow
opencv-python
transformers
torch
sentence-transformers
numpy
scikit-learn
flask
werkzeug